Create StreakLocalNotificationContentProvider
